requirements.txt
langchain_openai
langchain_core
langchain_community
streamlit
python-dotenv
langchain_ollama
langserve
fastapi
uvicorn
sse_starlette